Remove "Couldn't pair" prompt on ACTION_PAIRING_CANCEL
ACTION_PAIRING_CANCEL is triggered as a result of an ACL connection
going down, causing a "Couldn't pair" prompt. This can happen on some
car kits, even if the pairing is successful and the car kit briefly
tearing down the ACL connection after pairing and resuming it
immediately afterwards.

If a connection gets disconnected during a normal bonding procedure, the
UNBOND_REASON_REMOTE_DEVICE_DOWN reason code for the
ACTION_BOND_STATE_CHANGED already indicates this issue and will inform
the user.

Bug: 63975046
Test: manual
Change-Id: I663d8b7700741927c72e38e30cd993c0b81511a9

Former-commit-id: 61ad5ae921d75d0f72e8477ce224a1f7562ae81d