Ensure we leave the orientation changing state.
During the screen rotation animation, we cover the screen in a static
snapshot. Since failing to exit this state is somewhat catastrophic
we have a timeout which is intended to be a catch call. The timeout
works by clearing the mOrientationChanging state of all windows
and then triggering a layout pass. Recently however we made changes
allowing the orientation changing state to persist past the mOrientationChanging
variable. In the bug, we see many notices of timed out draw reporting, but
somehow we still haven't left the orientation changing state
and the screenshot is still visible. I have to admit...this is sort of a guess
I don't understand why we never reported the configuration to the client...though
it seems probably some intracicies of visibility could be at play.

Bug: 64127238
Test: go/wm-smoke
Change-Id: I5dea9bad29e7cdb43aa5e8c9ece5b0a86a1c8a09

Former-commit-id: 9c1c3a016a1d919e59721ef0eff8dbc67b61fb2c