Don't set inset frame for letterboxed app in fullscreen
When transitioning from split-screen to fullscreen for
letterboxed apps we were not clearing window inset frame
in some cases. This lead to incorrect mFrame calculation
and app looked like it was stuck in split screen.

This CL always clears inset frame for letterboxed app -
when task is fullscreen but app window bounds are not
empty.

Fixes: 64369819
Test: com.android.server.wm.WindowFrameTests#testLayoutLetterboxedWindow
Change-Id: I84d49642efb874caa55f2dbd8092d33e6b6f0ed4

Former-commit-id: c24f373d54ae6f9aee28380eb95fd0a97f0bd0b5