Document existing immutable behavior of ArrayAdapter
Various methods on ArrayAdapter throw exceptions if run on an adapter
created with one of the varargs constructors. This is because those contructors
create an Arrays.ArrayList, which does not allow modification. It is not
worth changing the existing behavior, but it is worth making the behavior
clearer in the docs.

Bug: 63802412  ArrayAdapter crashes when calling clear() directly after constructor
Test: docs-only fix. Verified with unit test that the docs apply correctly to the
newly-documented methods. Built the docs target successfully.

Change-Id: I75ae233879cc7db10853f93d78625a00675594b2

Former-commit-id: b7c6a5d7e91b39074428b39e918e092768593700