FingerprintController: Make sure to reset the fingerprint mode when going to sleep
Fixes an issue where the fingerprint controller could be stuck indefinitely, causing
the status bar to be in the SHADE state instead of the KEYGUARD state.
Now reset when going to sleep.

Also adds logging.

May also fix a second bug where the scrims were stuck in a bad state.

Change-Id: I55d456c5d157f92758267f684de2bd216ef07d94
Fixes: 64598807
Bug: 64594986
Test: mp sysuig; unlock with fingerprint; verify transition still works

Former-commit-id: ba7ca590afe9d99d93e6b81b0185b4b349476091