Add some preloaded classes related files and tools
Added copyright-header to be used by generation scripts to
automatically add a copyright header to preloaded classes.

Add a default blacklist file to avoid needing to manually
remove classes that can't be preloaded in the zygote.

Added a script go generate a preloaded classes list from a raw class
list. This script adds the copyright headers and comments and removes
blacklisted classes.

Bug: 37966211
Test: manually run the scripts

Change-Id: I166f990c41978c80134d2f1e12c1b1079a12752a

Former-commit-id: 678ba75bb9b700a927d2f8f3bea725086c59405d