USB debug dialog won't give permission if obscured.
Previously, the USB debugging dialog was suceptible
to tapjacking over tcp if a malicious party obscured
the USB debugging text to mask the id of the system
requesting permissions.

The fix prevents users from giving permissions if the
dialog being displayed is partially obscured.  Instead,
it will present a toast explaining to the user why they
cannot give permissions.

Test: manual
Bug: 62187985
Change-Id: I3bdcd1876cd6dbe8a728bbce74edb52ab79f3e4c

Former-commit-id: 830d54f2c75bee8bda4c3bdd3d62c9f2f58934b7