Improve alarm in-flight diagnostics
Also, use .equals() rather than == when searching the list of in-flight
operations.  Sometimes we appear to fail to match PendingIntents when
their send() completes, for unknown reasons, and this specific point of
usage is one where a mismatch might lead to the observed behavior.
(It is still unknown why == might ever fail here, and the problem is not
readily reproducible.)

Bug: 64358454
Test: manual, observing correctness of the diagnostics

Change-Id: I4f5c07b0df44dfb4c292976b3a49482ed408d12c

Former-commit-id: eabba7366e057fb5c103d5aa35f497a518f78672