Adapt ICU 59 API update to Doclava changes.
Add methods that were missing from the previous API update because the
update-api version used for that had different behaviour from the
current one.

Specifically overrides for methods that differ only in sychronization
are now considered API-worthy and weren't before (b/62576297).

All methods added in this commit are methods that exist in the base
class without synchronization and are synchronized in the class they are
now being added in.

Bug: 64807220
Test: m checkapi
Change-Id: Iff584fb72dc8d651f9482b573f4b8cec89840ca6

Former-commit-id: 0f1314db8ffbd5d2fbe104220f25b448e13a62ad