Add self update permission
This permission is a subset of INSTALL_PACKAGES that only allows a
package to update itself. Many of the users of INSTALL_PACKAGE claim to
be using it solely for self updating but INSTALL_PACKAGES gives a lot
more power than is needed. This limits those priv apps to what they
intend to do.

Test: manual
Change-Id: I3442ad1a285df660b8c2635e3db59d7bc0662324

Former-commit-id: cfdc1eeacf6e84da919772d95c26d3a0b31a3ae9