Use the installer package name for broadcasts
We would send the broadcast to the installer of record for
the _removed_ package [ie. replaced]. There was no value
containing the installer of record for the _installed_
package. Now prefer to send the broadcast to the installer
of record for the install. Otherwise, use the installer of
record for removal.

Change-Id: Idb1a7fa522a19273eb3641746e1966aa7af97e7c
Fixes: 62905206
Test: Manual. 'adb install test.apk' doesn't fire broadcast to the installer.
Test: Manual. 'adb install -i com.foo test.apk' fires broadcast to "com.foo".
Test: Manual. 'adb install -r test.apk' fires broadcast to "com.foo" installer.
Test: Manual. 'adb install -r -i com.foo2 test.apk' fires broadcast to "com.foo2".

Former-commit-id: 87505bb58a49768703fde504b7eb4a40429d7721