Reset functions to null when user changes.
When MTP or PTP is in the config and user changes,
it clears dataUnlocked but sets the functions to
mCurrentFunctions which will leave MTP or PTP in the
config. If dataUnlocked is false, we want to remove
MTP/PTP from the config when adb is enabled, and leave
them in otherwise. This is done automatically by
setting functions to null.

Bug: 64822515
Test: Change users while adb  / mtp connected.
Change-Id: I82561f82e06f247c967438a473b08b5098d39ccb

Former-commit-id: e537a8313987bd10208281c4fa5c6834b4fd1d15