Check correct DisplayContent when making window freezing screen
When an AppWindowToken is made visible, it checks if any of its windows
are in a state that necessitates freezing the screen. Windows should
check whether their DisplayContent is ready, not whether the default
display is ready.

For apps on secondary displays, the incorrect check was causing spurious
orientation changes for their starting window if the primary display was
off, which was interrupting the task transition animation.

Bug: 34280365
Test: manual
Change-Id: Ifd3be2aedc32c42f3c3f1b15f9f0a69a839f8c62

Former-commit-id: 5a9ecee75b9230218f012d41ee137724e03c01e5