Reset update spec after canceled layout
When we cancel a layout there are two cases. Either we already scheduled
the next layout. In this case operate as before as the update-spec was
updated to the new update.

In there is no new layout we have to reset the update spec as the last
command was canceled and the state of the remotePrintDocument has
become invalid. The resetted update spec forces a new update the next
time update() is called.

Fixes: 64911776
Test: cts-tradefed run cts-dev -m Print
      Added two new tests reproducing the scenarios mentioned in the
      bug.

Change-Id: Icad8da5fd985aed1cc201076ee7262f6d6541380

Former-commit-id: 196784ac1da21eac2c6801484a68fb6b1718765e