Fix stuck activity
Assume the following sequence of events:

- prepareAppTransition
- setVisibility(app)
- okToAnimate -> false
- setVisibility(app)

In this case, the first setVisibility call sets a dummy animation
but the second setVisibility call removes the app from
mOpeningApps but does not clear out the animation. In that case,
the app is stuck with the dummy animation set that will get never
cleared.

To fix this, we always reset the animation if a dummy animation is
set when changing the visibility, no matter we use a transit or
not.

Test: Unlock with fingerprint from AOD to homescreen
Change-Id: I0dbe2d9836238aba7f16d3e45649fd4b82ff52f6
Fixes: 65067824

Former-commit-id: 4d1835db5d98f75bbd2dd1e8bed3a06d7f1ccc46