resolve stage dir before validation
We rely on the stage directory being correctly resolved
in order for validation to occur. This forcably happens
when commiting [after validation], reading or writing
to the session. For normal installs, this isn't an issue
since the session will be written to, thus, resolving
the stage directory.

Instead of waiting until the commit, resolve the stage
directory just prior to package validation.

Change-Id: I667b9d27749405d24e3a3c6666e9591ce3e883df
Fixes: 65369970
Test: bit FrameworksCoreTests:android.content.pm.PackageManagerTests\#testInstallNonexistentFile

Former-commit-id: 6de494dfb36ba5677996c844198210fe392a1ba7