Fix issue #65359137: Apps with target API 26 can bypass background...
...limit introduced in Oreo by swiping away from the task list

There was an edge case where, when restarting a service, if the
process wasn't already running we would bring it up directly in
the idle state and never go through the mechanism of force stopping
services once an app goes idle.

The fix for now is pretty targeted, detecting when we have a new
uid that is immediately idle, and in that case doing the same
thing we do when an existing uid goes idle (stopping any services
in it).  This means that in this flow you will get a glitch of
the service starting and then immediately stopping when idle is
applied at that point, but we do end up ultimately in the correct
final state.

Test: manual
Bug: 65359137

Change-Id: I64c29975ad75e7d7390512bf59673d423df16cf2

Former-commit-id: a857e65c53957f87036b7962750bbdece6012b53