Limit setActivityType safety check to system process.
WindowConfiguration.setActivityType() has a safety check to help
developers avoid logic errors that can cause an activity or its
parent container to change activity type which can be dangerous.
However, this protection is mainly there for code running in the
system server and shouldn't affect coding running in the client
process. For example, we don't want to crash code running in the
client process because of the way the app decides to combine
various configurations they get from the system. This change limit
the check to just code running in the system process.

Change-Id: Ie315d8d69b5ac03e4fd4ea1e3353a37d097af96d
Fixes: 65403289
Test: bit FrameworksServicesTests:com.android.server.wm.WindowConfigurationTests

Former-commit-id: 6cbbc9a0a13caa1e9f748e1d2a3d1fbf83c1407c