Enable new initial-focus behavior for P
This means that, by default, nothing gets initial focus
in touch-mode.

In order to have something focused in touch mode at
activity start, an app must explicitly provide a
<requestFocus /> tag. Since this tag uses the "default"
focus mechanism, it can be placed in a parent and focus
will go to whatever view is the default focusable in
that hierarchy.

This takes effect for targetApi >= P

Bug: 34520588
Test: Added CTS tests in View_InitialFocusTest
Change-Id: I74839bb86079558eaa5c96de52220a7d2c47d901

Former-commit-id: bdc66cb5a0ef513f4306edf9156cc978b08e06e4