Fix light status bar when transitioning home -> recents
In the home -> recents animation home is exiting but recents
is still translucent because this is how the animation is
designed. In that case, mWinAnimator.mShownAlpha is 0 such that
mTopFullscreenOpaque becomes null, creating a flicker if both
home and recents have light status bar flag set.

We fix this by refining this condition: It was added such that
a starting window for an activity that might never become visible
does not affect the systemUI flags. Now, we only avoid affecting
the flags if a dummy animation is set (a transition is about to
start) and shown alpha is 0, which indicates that the window is
entering.

Test: go/wm-smoke
Test: With light wallpaper, go home -> recents
Test: All other common transitions with light status bar
Test: Open camera, go to gallery, go home, double click power
button to reopen camera (regression test)

Change-Id: I6cddc3b8eb09c547618d7572195889de7766d838
Fixes: 65398267

Former-commit-id: b7202a8339652a721a285c37358e38b4926f8e25