StatusBar: add DISABLE2_SYSTEM_ICONS flag
Add a StatusBarManager flag to control the visibility of system icons
such as connectivity info, battery, etc. Note that the clock is
controlled separately through DISABLE_CLOCK.

Also note that this flag differs from DISABLE_SYSTEM_INFO, which
controls the system icons and the clock as a whole, but only works for
the collapsed status bar (CollapsedStatusBarFragment).

Bug: 65813398
Test: runtest systemui
Test: manual, by using a modified version of TestDPC and
      DevicePolicyManagerService

Change-Id: Icb61c7bb1a625edd818be7a66d7f15c3e812237b

Former-commit-id: 6a79b0dff1d82e51d5f0a1d30f8491b6dcb4d6f5