Make sure wm Stack gets override config on creation
Bug was introduced in ag/2840516 where we started to use
ConfigurationContainerListener to send override configuration changes
from AM containers to WM containers. However, the stack object on the wm
side didn't set its controller in the constructor like the other
container objects so will not have the right information for anything it
tries to do in the ctor or shortly after.
We now pass the controller into the container ctor like we do for the
other container classes.

Change-Id: Iae4166a3511e69f4823e7fb9191774c34cc02700
Fixes: 65738339
Test: run-test CtsWindowManagerDeviceTestCases android.server.wm.CrossAppDragAndDropTests

Former-commit-id: 704a3c0d667049753fe58716db58099bec1768c2