Remove obsolete & broken handling of precompiled sys apps
During the times we used verify-profile for OTAs we had to ensure that
prebuilds do not regress by verifying at runtime. For that we added a
hackish check in PMS.

Now, we no longer use verify-profile and we have no configuration which
will leave the prebuilds unverified (assume-verify does not count here).
So, we can remove this check which blocks system apps to be compiled with
speed-profile.

(Also worth mentioning that
!Environment.getReferenceProfile(pkg.packageName).exists() was broken...
most likely because of an SElinux policy update which we did not handled
:/ )

Bug: 66903791
Test: build & boot
Change-Id: I981e463d61c42ab8455d489ff86cc300632a2a90

Former-commit-id: efc022768eba2f33d3fadbda7eaa6adf1730d3fc