Fix launch of activities into VR Virtual display
When launching a 2D activity in VR mode by the VR app, we want to launch
it in the virtual display created by VR. But, this fails since the
virtual display owner is system and the calling UID doesn't have
ACTIVITY_EMBED permission.

We address this by making the virtual display public since it's lifetime
is only VR mode and we want to allow all apps to launch 2D activities
and route them to the vr virtual display.

Bug: 66475142
Test: Launch VrSettingsActivity by clicking on gear and check "adb shell
am stack list" to verify it launched on the virtual display

Change-Id: Idde63ea63c953c6cf3d4ea46989a963a5651ce8a
Signed-off-by: Karthik Ravi Shankar <karthikrs@google.com>

Former-commit-id: 4c6573889567dbe2dc0fad6217818aab6b7a7a1f