Decouple Content Name Locale from Content Language Locale
There was an assumption that content language Locale was
a 1:1 mapping with content locale name, which is not the
case. This change separates the two and provides a new
API for retreiving the set of Locales for content name.

Test: Manual
Bug: 67112896
Merged-In: If87094f15e3a1dfe36d37745ed3d2cf79cc441f1
Change-Id: I44cb527ceb77fe321500f9d5fc00f6880ee52cb7

Former-commit-id: a722139f9cdf3f3368589cf5667b29e26bd29459