Insert org.apache.http.legacy at the start of the shared library list
Ensure consistent behaviour depending on whether OAHL is on the
bootclasspath or not.

When OAHL is on the bootclasspath the search order is (where ... is the
other libraries on the bootclasspath):
   OAHL
   ...
   shared libraries
   optional shared libraries
   APK

Prior to this change the OAHL was added to the end of the shared
library list which meant the search order (when OAHL is not on the
bootclasspath) would be:
   ...
   shared libraries
   OAHL
   optional shared libraries
   APK

After this change the order will be:
   ...
   OAHL
   shared libraries
   optional shared libraries
   APK

The slight difference at the beginning is not an issue because there are
no conflicting resources or class files between OAHL and the other boot
libraries.

Bug: 65552462
Bug: 18027885
Test: build, flash, check systrace when starting GoogleDialer
Change-Id: Ifcb4d50c13e35eebac4d18f8f0f10dd0734e8896

Former-commit-id: 035820cbda4c208bec8129d0e1e43f986d2dff87