Throw OOME if Bitmap.nativeCreate fails
Bug:33846679
Test: I6ab6cb7a3b3151641a9f9b02b0bfc484e0a4524b

This matches the old behavior, prior to switching from Java allocated
pixel memory to native allocations (b/27762775). It also better matches
what has happened - we ran out of memory. (Better than the current
behavior - NullPointerException in the Java code, or an alternative
solution of returning null, which would likely result in NPEs in the
calling code.)

Change-Id: I3958ed1106ac94fb1d3f30e044b620d984875211

Former-commit-id: f3a0299753dffe3cd79d62b99001754fe6dbf524