Add keyguard state as a gate for VR Mode.
Problem: When the user toggles power button while in VR, they will be
taken to a black VR screen that occludes the keyguard.

VRManager previously used wakefulness and screen off as gates for VR Mode. However,
the keyguard state can be on or off when the device is awake so we need
to listen to the keyguard state as well.

Change:
1) Have ActivityManagerService tell VrManagerService about the keyguard
state the same as it does with wakefulness.
2) Stop applying any pending states when the device comes out of
sleep/screen-off/keyguard to force user into putting-headset-on (DON) UX
in those situations.

Bug: 66696304
Test: Manual - toggle device's power button and see logs of keyguard.
Test with all keyguard types.

Change-Id: I4ae03cb1c8c13acdc8cd11839f178bc8d8f6f515

Former-commit-id: b5cbecb16a53e74edd626b1571a47aa261482efd