Fix parcelling of Kernel Wakelock/reason
In BatteryStatsImpl.java:
Previously, if inclUid == false, then kernelWakelockStats and
wakeupReasonStats were not printed to the parcel, but only a single 0
was written in their place. In the parcel-reading code, their sizes are
read, so if they are not present, there should be a 0 for each of them.
In other words, two 0s should be written if they are not present,
instead of just 1.

Test: none
Change-Id: I1d443e9659646f0d30063f275a6a9db42b0a3906

Former-commit-id: ba78b2bcc73565b68394764ef8a302df64d60fc5