Add CarrierConfig for Invalid Call Forwarding number
In some situations, an invalid Call Forwarding number
from a carrier's network can cause the modem to send back
invalid characters instead of a number.

This change adds a carrier config option to replace the invalid
text with "Voicemail" in the Call Forwarding settings when
it is received from the network.

Bug: 67267077
Test: Manual
Change-Id: I26f9b80487f92340d1df2cf22bb5a616edcbdebf

Former-commit-id: 110856431341a221206b29e25fb4aa986da662da