Expose SmsMessage#FORMAT_ constants as public APIs.
These constants are intended to be arguments for
SmsManager#injectSmsPdu which is already public, and they are
referenced from CTS as well. Therefore, they should also be public.

Also add a @StringDef for these values and clean up the Javadoc while
we're here to use properly-linked references to other APIs.

Change-Id: I4e513f49f18350b93cccdc0d8a5847d31eefe7e2
Fixes: 67751433
Test: TreeHugger (no functional changes)

Former-commit-id: beb90fda29d1b6d8cf3a09468b410aad3527f280