Fixes retrieving fullscreen stack when recents is not available yet
First time going to recents would not have the recents stack created yet
and so SystemUI would fallback to getting the window rect from the
fullscreen stack. The code getting the fullscreen stack passed an
undefined window mode which would return a null stack. Fix this by
asking for a fullscreen window mode.

Change-Id: I372bcf5ccffd87af467a8a1a8cee46def1640654
Fixes: 67864447
Test: manual - on boot, launch app, press recents button

Former-commit-id: 30aadfdc474f36943444e29fabef51646ca769b0