Prevent touch delegate from blocking parent events
Make sure that TouchDelegate does not interfere with
handling of parent's events, such as onClick().
Currently, the variable mDelegateTargeted is sticky,
and after the TouchDelegate is first activated,
stays true for the remainder of its life. Therefore, the
parent view will no longer receive clicks after the
TouchDelegate is first clicked. This patch forces the
re-evaluation of mDelegateTargeted on each new ACTION_DOWN
event.

Test: bit CtsViewTestCases:.TouchDelegateTest
Bug: 65392236
Change-Id: Ifa6430ce6e8400c0411df7da66281219ea97733c

Former-commit-id: d27a6319f21800cf423e0a7d3d49acc88eaa38ed