docs: Clarified description of height and width attrs
The existing description had said the attributes specified the exact
height/width *in pixels*, but (as clicking through on the links
showed) the height/width could be specified in a number of units,
including inches and millimeters.

Rephrased to be more accurately vague ("specifies the exact height/
width"), and if people click the link they get the details.

See first comment for doc stage location.

Bug: 63614266
Test: make ds-docs
Change-Id: I9a2873683d19bcfbb9f2c4a603438d858539b650

Former-commit-id: e493a145af2f54c3e15d1360b05948cf1fc84c95