Sets up ContextHub service/manager interface for transactions
Creates a IContextHubTransactionCallback, a callback that can be invoked
at the service as a proxy after completion of each transaction. The
ContextHubManager generates the callback for every transaction, and
passes it down to the service. This removes the need for the transaction
ID to be tracked at the client application side.

Bug: 67734082
Test: Make from root
Change-Id: I89a77a308671e7a84e50b27eb9c4de012dedc310

Former-commit-id: 6c37fd0f538c4a6f40589ba1e91d95449f0b8195