Remove libmedia_omx_defaults from libmedia
libmedia_omx_defaults includes source files and this default was
used in both libmedia.so and libmedia_omx.so. This is an issue because
libmedia_omx and libmedia are both loaded in the framework process
through dependency chains such as libstagefright ->
libstagefright_omx_utils -> libmedia_omx.

To fix the issue, libmedia_omx_defaults is removed from libmedia and
libmedia now depends on libmedia_omx. Other modules/libs that originally
depended on libmedia and need access to libmedia_omx_defaults' source
files, now depend on libmedia_omx as well.

Bug: 67871061
Test: mma -j and boot test
Change-Id: I096d58d1e3338b70d5dae9d0f9919b908070d071

Former-commit-id: a0f51ba4845194baa7b1ce4d15888acc2267501b