Don't drop system server StrictMode violations.
The system server doesn't have a ProcessRecord, and all the
downstream already checks if it's null.  (Someone incorrectly added
a bailout-when-null check.)

We've actively burned down the number of system violations to a
minimum, and we're chasing the remaining violations with the intent
of them staying at 0, meaning we're okay to remove the batching
logic.  (The output format of batched violations is tricky for
dashboards to parse.)

Test: cts-tradefed run commandAndExit cts-dev -m CtsOsTestCases -t android.os.cts.StrictModeTest
Bug: 68662870
Change-Id: Ic1b81200c9550a316939e890a8cea8f37cfbb30b

Former-commit-id: 2bccbc8ec1c3177d3db5bd71d24855ab212ee5a9