If the absolute value of a sweep angle >= 360 then draw it as an oval.
Skia's drawArc API is defined to keep winding when drawing an arc
with a sweep angle > 360.  This allows for special path effects
and for points on the arc to potentially be double drawn. This is
not the case for the HWUI implementation that optimistically turns
the call into a drawOval.  This CL updates Skia's pipeline to be
in line with the HWUI implemenation.

Test: CtsGraphicsTestCases were added to cover this
Bug: 68764873
Change-Id: I22cd15a53cd2df44421518256d45a2b5d0525a86

Former-commit-id: ef3b2187354bd648bf9e1ab1d0757f5de22a42ce