Bypass a11y cache when requested
When refreshing a node from the cache, we would still ask
the cache to update itself in the middle of the refresh,
which led to nodes being put in inconsistent states.

Also rolling back ag/3051037, which was a temporary fix for
one crashing bug.

I'm also doing a bit of cleanup. Several methods in
AccessibilityInteractionClient should really be static,
since they only manipulate global state.

Also, since I was already adding tests, I'm moving the
tests for android.view.accessibility to their proper
place in the source tree.

Fixes: 68987775
Fixes: 67735115
Test: Adding a test to verify that we don't touch the cache
when refreshing a node. This test fails without the change
to AccessibilityInteractionClient, and passes now.

Change-Id: If642ca79bc414ef9846b7b2a2277870f4143cb98

Former-commit-id: c140fdc34263caaa021d5c4ec438384de484420d