Update Roboletric to 3.5.1
To be able to migrate Robolectric and bump SDK we have to load our classes from
the system class loader instead of the android jars. This wasn't a
problem before because the classes weren't in the jars correspondent to the old
SDK that the tests were using. To be able to do this I created a custom test
runner and a custom annotation to be used in tests to load the specified
classes from the system class loader. Check comments on file. Also, made changes
to be compatible with new Robolectric version and bumped SDK.

Bug: 69254249
Test: m -j RunFrameworksServicesRoboTests
Change-Id: I538ef70cfae8ebeea047f275b9920cb2e814f4da

Former-commit-id: 90a880232847fd7d963c5805c9b76db66df115d8