Move home stack behind top stack when split-screen is dismissed
Whenever split-screen is dismissed we were returning to whatever stack
was on the other side since that is the next thing in terms of z-order.
However, visually to the user we should be returning home. This CL makes
it so.

Also,
- Fixed issue were FLAG_ACTIVITY_LAUNCH_ADJACENT might not work due to
early return in ActivityStarter.getLaunchStack since the stack return by
mSupervisor.getLaunchStack() is never null.
- Just set windowing mode when AMS.setTaskWindowingMode is called vs.
creating a stack to re-parent to.
- Fixed issue with reporting multi-window mode changed to be based on
onConfigurationChanged() for the activity record.
- Corrected a few issues with order of sending pip mode changed and
multi-window mode changed due to now sending multi-window mode
changed in onConfigurationChanged().
Change-Id: I4893937853968bfd0b437dc646a82ead8f127f2e
Fixes: 69378933
Fixes: 68953884
Test: ActivityManagerSplitScreenTests.onSplitScreenModeDismissed

Former-commit-id: eb76b762e2c2e6e02a4d7456e16fb5cf9881f342