Correct A2DP status when audio routes changed
When a Bluetooth device is connected and then a media is started,
sometimes MediaRouterService gets media playback status change first
and get the Bluetooth device status change later. This causes the
media plays through the phone speaker instead of BT devices.

This CL prevent that situation by not changing A2DP status when
BT device is not connected, and by calling restoreBluetoothA2dp
at the end of dispatchAudioRoutesChanged.

Bug: 69499034
Test: manually tested the bug, passed Media CTS
Change-Id: I7547a39e5ce58f58c249a25efa55446a3b023416

Former-commit-id: 90f4c8959e6bff9825c033966d461ed0d0e3b6cb