There is only one Generic.kl.
Generic.kl is a special file that defines the default mappings for all
devices. There can only be one and it cannot change regardless of the
product area the device is in. This should *never* have been checked in,
especially not without someone on the input team looking at it.

Also, add support for the POUND and STAR keys, which is why it was
forked in the first place.

Test: N/A
Change-Id: I894cbefe20fc77d827dc500daa3a380f252eb8b6

Former-commit-id: 4cb42f3894066d4bc54417501354d782b3bcfeee