Changed Dataset.setValue() methods to allow filtering with authentication.
The Dataset.setValue() methods were created before Datasets supported
RemoteViews as presentation. As such, the service always had to provide a
value for the dataset fields, and the value could not be used for filtering
(otherwise the user could "guess" what that authenticaed values were).

But once Dataset required RemoteViews for presentation, the field values were
not used anymore when the dataset required authentication (in fact, the
setValue() methods accepted null for the value), so now we can always use the
field values for filtering.

Test: atest CtsAutoFillServiceTestCases
Fixes: 69458842

Change-Id: Ia58af8bb946f31c3806c28570b2445870d7ccf0c

Former-commit-id: bbf8549ad0bab9f73501956278b25c57a17eb5ab